Fix storage slot value removal issue (0.127) (#10945)
Fix storage slot value removal issue (#10939)

This PR fixes storage slot value removal in modularized flow.

When the account is a contract we need to populate `contractKvPairsNumber` base on its storage values.

Currently it is always 0 and when we try to make a storage slot update where we set value to 0 we get an error.

`Account` - set contractKvPairsNumber
`AccountReadableKVState` - set `contractKvPairsNumber` to the `Account` model
`StorageContractTest` - add tests for the previously failing scenarios

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>