Invalid sender bug fix cherry pick (0.127) (#10946)
* Add exception handling for invalid sender (#10936)

Invalid sender leads to 500 internal server error, because of pre-handle check failure in services, which is not properly handled in the StandaloneDispatchFactory, and as a result null pointer is returned to the mirror node. The current pr is workaround to the issue that is present in services.

---------

Signed-off-by: sdimitrov9 <stoyan.dimitrov@limechain.tech>
Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>

* Add `isModularizedCall` exception parameter in the mono flow (#10893)

In the Grafana logs there are some exceptions from the mono code where the isModularizedCall is passed like null instead of false. This PR sets the correct value.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>
Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>

---------

Signed-off-by: sdimitrov9 <stoyan.dimitrov@limechain.tech>
Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>
Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>
Co-authored-by: Stoyan Dimitrov <stoyan.dimitrov@limechain.tech>
Co-authored-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>