Add the missing contract key in the state (#10968)
This PR is a workaround for the following scenario: a contract with a null key was deployed and imported in the mirror node DB (before the fix in hiero-ledger/hiero-consensus-node#15990). The parent transaction has a payer which is the one set as a sender in the transaction. The key verification is fine. A child transaction is created from the contract call. The latter transaction has a payer set to the contract and its key is verified against the contract's key which is null, hence the error in this issue.

This was resolved in services by adding a default key from ContractID(shard, realm, num). This PR in the mirror node mocks this logic until we have a real DB migration - when an account is read in the AccountReadableKVState its key is set before the account is put in the state as an object. If the key is null and the entity is of type 'contract' now a dummy key is added from ContractID(shard, realm, num) as if the migration was already run.

Note: the contracts following the fix in services do not have this issue.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>