Fix `NullPointerException` on missing account key (0.128) (#11006)
Fix `NullPointerException` on missing account key (#11000)

In the consensus node there are no missing account keys in the state and if there are this is treated as a catastrophic failure. In a previous PR we added a default key for contracts. In this PR we add a default key for accounts. With this change we should be in sync with the consensus node. An integration test was added, as well as 2 junit tests that verify that the default account/contract keys are returned as expected.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>