Modify findByIdAndTimestamp query (#11005)
It was observed that the findByIdAndTimestamp query tends to execute slower that expected, so this PR proposes a query optimization. In the original query the left join joins the entire token table to the subquery using the token_id.
The optimizer has to evaluate the whole table or at least use an index scan on token_id. The larger the token table, the larger in impact on the performance. In the new solution only 1 row is joined to the subquery.

Signed-off-by: sdimitrov9 <stoyan.dimitrov@limechain.tech>