Adapt state to check for new state keys in hedera-app (#11004)
This PR introduces helper methods and a StateKeyRegistry class that lists all supported state and singleton keys for the current hedera-app version. During schema initialization, we now validate that all keys are known.

If hedera-app is upgraded and introduces new state keys that aren't present in StateKeyRegistry or missing in DefaultSingleton list of keys, an UnsupportedOperationException will be thrown. This helps catch unimplemented state logic early when bumping hedera-app versions.

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>