Fix `NullPointerException` on get token treasury account (0.128) (#11055)
Fix `NullPointerException` on get token treasury account (#11053)

This PR changes the way the token treasury and auto renew accounts are added to the state. Now they are added only with account nums instead of account aliases. This is done in order to be compatible and in sync with the hedera.app dependency and the consensus node.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>