Prevent `null` singleton values (#12402)
In hedera-app bump from 0.64.3 to 0.65.0 the WritableSingletonStateBase class was changed to abstract. Also, its get method was modified so that the current singleton value is returned even if it is null, which was not allowed in the previous release. Also, with this hedera-app bump the NullPointerException issues started to appear. This PR returns the previous behaviour - return the modified value only if it is not null, delegate to the super implementation (our custom singleton states that return non-null values in get()) if null.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>