Bump hedera-app to 0.69.0-rc.2 (#12533)
This PR bumps hedera-app dependency to 0.69.0-rc.2. In addition, it removes the previously copied files from upstream when we bumped hedera-app to 0.68.0.

Unfortunately, it seems there is a missing configuration in the module definition in hedera-app and it was required to add a temporary dependency for implementation("com.hedera.hashgraph:app-service-entity-id-impl"). I will open a ticket upstream to resolve this issue.

---------

Signed-off-by: Ivan Kavaldzhiev <ivankavaldzhiev@gmail.com>